isolation feature: disconnect EDB from power source
To let the APP access the power source exclusively.

Note that the isolation is not complete when Vin is higher than Vedbmcu
because the P-fet is not fully closed in that case. Potentially fixable
by adding an N-fet at the gate of the P-fet. But that is OK, because EDB
gets power only if there is abundant power (because otherwise the app
would droop the power source well below Vedbmcu).