Revert "decode: print edit distance to beacon"
This reverts commit 0ec9888137f6bdca9c5f07d0e6527888d774fc9c.

Does not make sense. The codeword is:
preamble+parity_byte+data_byte+post_amble

So, if the receiver prints byte XX that implies that the waveform
matched (above a threshold) to the code word .. .. XX ..  So, there's
not such thing as one bit of the data byte flipping in transit --
because then the waveform would change such that it matches an entirely
differenty codeword (with data byte YY).

This also implies that any error correction (if it were added)
would need to be for multi-byte packets, and it would need to
be able to correct in multiples of 8 errors.

Correction at the level of the one data byte does not make sense.

Checksum at level of the one data byte might be sort of useful
in that it resticts the possible bytes that could be sent. But,
it's a shaky benefit: if you get a byte with a bad checksum
(i..e a byte outside the set of possible bytes that could have
been sent), then you can throw it away as noise, but if you
get a byte with a passing checksum, you don't really know if
it was actually sent or if it was noise. Even the probability
of noise-or-not-noise does not change by adding the checksum,
because ALL bytes have equal probabiilty of being generated
as a result of noise -- the noise creates bytes without caring
about your special set of bytes.

With a checksum on multiple bytes of payload, the probability does
change, because the probabiity that a valid sequence of bytes was a
result of noise multiplies with each byte, dropping.